+++ public/Projects/Anubhav/Weather App/code/templates/partials/footer.hbs
@@ -1,0 +1,6 @@
+<footer>
+    <p>
+        Created with ❤️ by Anubhav
+    </p>
+
+</footer>
