+++ public/Projects/Anubhav/Weather App/code/templates/views/index.hbs
@@ -1,0 +1,52 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+   {{>header}}
+</head>
+
+<body>
+    {{>navbar}}
+    <div class="container-fluid main_header">
+        <div class="row">
+            <div class="col-md-10 col-12 mx-auto">
+                <div class="row">
+                    <!-- Left Side -->
+                   <div class = "col-md-6 col-12 main_header_left">
+                       <p>Weather Web App!!</p>
+                       <h1>Seach Your City and <span class = "text_clr">Find the Latest Weather Here</span></h1>
+                       <a href="/weather"><button>Check Now!</button></a>
+                   </div>
+                    <!--Right side  -->
+                   <div class="col-md-6 col-12 main_header_right">
+                    <div id="carouselExampleControls" class="carousel slide" data-ride="carousel">
+                        <div class="carousel-inner">
+                          <div class="carousel-item active">
+                            <img class=" w-100" src="images/7.png" alt="First slide">
+                          </div>
+                          <div class="carousel-item">
+                            <img class=" w-100" src="images/3.png" alt="Second slide">
+                          </div>
+                          <div class="carousel-item">
+                            <img class=" w-100" src="images/4.png" alt="Third slide">
+                          </div>
+                        </div>
+                        <a class="carousel-control-prev" href="#carouselExampleControls" role="button" data-slide="prev">
+                          <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+                          <span class="sr-only">Previous</span>
+                        </a>
+                        <a class="carousel-control-next" href="#carouselExampleControls" role="button" data-slide="next">
+                          <span class="carousel-control-next-icon" aria-hidden="true"></span>
+                          <span class="sr-only">Next</span>
+                        </a>
+                      </div>
+                    </div>
+                </div>    
+                </div>
+            </div>
+        </div>
+    </div>   
+    {{>footer}}
+</body>
+
+</html>
