+++ public/Projects/Anubhav/Weather App/code/templates/views/weather.hbs
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+   {{>header}}
+</head>
+
+<body>
+    {{>navbar}}
+    <div class="container-fluid main_header">
+        <div class="row">
+            <div class="col-md-10 col-12 mx-auto">
+                
+               <div class = "main_content">
+                   <form class="temp_form">
+                       <input type="text" id="cityName" placeholder="Enter your city name" autocomplete="off">
+                       <input type="submit" value = "search" id="submitBtn"> 
+                   </form>
+
+               </div>
+               <div class = "tempinformation">
+                   <div class="top_layer">
+                       <p id = "day">Sudnay</p>
+                       <p id="today_data">26 SEPT</p>
+                   </div>
+                   <div class = "main_layer">
+                       <p id = "city_name">Output Here</p>
+                       <div class = "middle_layer data_hide">
+                           <p id = "temp"><span>0</span><sup>o</sup>C</p>
+                           <p id = "temp_status"><i class="fa fa-cloud" aria-hidden="true"></i></p>
+
+                       </div>
+
+                   </div>
+
+               </div>
+                
+            </div>
+        </div>
+    </div>   
+    {{>footer}}
+    <script src = "js/main.js"></script>
+</body>
+
+</html>
