--- public/Projects/Anubhav/Weather App/public/css/style.css
@@ -1,239 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Piazzolla:wght@900&display=swap');
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-    font-family: 'Piazzolla', serif;
-}
-
-:root{
-    --main-bg: #1e272e;
-    --main-color: #0097e6;
-}
-html{
-    font-size: 62.5%;
-}
-
-body{
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-}
-
-.row{
-  margin: 0;  
-}
-
-.main_menu{
-    background-color: var(--main-bg);
-}
-.navbar{
-    color: #fff;
-    background-color: var(--main-bg);
-    padding: 2rem 0;
-}
-.navbar li{
-    margin-left: 2rem;
-}
-
-.navbar a{
-    color: #fff;
-    font-size: 1.8rem;
-
-}
-.navbar a:hover{
-    color: var(--main-color);
-}
-
-.navbar-toggler{
-    background-color: gray;
-}
-
-.fa-cloud-meatball{
-    transition: all linear;
-    animation: rot 2s linear infinite;
-    color: var(--main-color);
-}
-
-@keyframes rot{
-    0%{
-        transform: rotate(0deg);
-    }
-    100%{
-         transform: rotate(360deg);
-    }
-}
-
-.main_header_left{
-    height: 70rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
-}
-.main_header_left p{
-    font-size: 2rem;
-}
-
-.main_header_left h1{
-    font-size: 4rem;
-    margin: 1rem 0 4rem 0;
-    text-transform: capitalize;
-}
-
-.main_header_left .text_clr{
-     color: var(--main-color);
-}
-
-.main_header_left button{
-   border: none;
-   padding: 0.8rem 1.8rem;
-   font-size: 1.4rem;
-   border-radius: 2rem;
-   text-transform: uppercase;
-   box-shadow: 0 8px 6px -6px #000;
-   outline: none;
-}
-.main_header_left button:hover{
-    box-shadow: 0 8px 6px -6px var(--main-color);
-}
-
-.main_header_right{
-    height: 70rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
-}
-.main_header_right figure img{
-    filter: drop-shadow(0 0.5rem 1rem rgba(0, 0, 0, 0.3));
-}
-
-.temp_form{
-    width: 60%;
-    position: relative;
-    margin: 5rem auto;
-}
-
-.temp_form input[type="text"]{
-    width: 100%;
-    padding: 2rem 5rem 2rem 2rem;
-    background: #1e202b;
-    color: white;
-    outline: none;
-    border: none;
-    font-size: 1.7rem;
-    border-radius: 3rem;
-}
-::placeholder{
-    color: #bfc1c8;
-}
-
-.temp_form input[type="submit"]{
-    position: absolute;
-    top: 5px;
-    right: 5px;
-    bottom: 5px;
-    border: none;
-    background: #009ad8;
-    padding: 0 3rem;
-    border-radius: 3rem;
-    color: white;
-    font-size: 1.6rem;
-
-}
-.temp_form input[type="submit"]:hover{
-    color: #009ad8;
-    background-color: #fff;
-
-}
-.tempInformation{
-    width: 40rem;
-    margin: auto;
-    min-height: 20rem;
-    background-color: #323544;
-    border-radius: 1rem;
-    padding: 1rem;
-    font-family: "Roboto", sans-serif;
-}
-
-.top_layer{
-    width: 40%;
-    margin: auto;
-    background-color: #323544;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    border-radius: 1rem;
-}
-.top_layer p{
-    color: #fff;
-    font-weight: 300;
-    padding: 2rem;
-    font-size: 1.3rem;
-}
-.main_layer{
-    width: 100%;
-    height: 80%;
-    display: flex;
-    flex-direction: row;
-    border: #0097e6 solid;
-    border-radius: 80px;
-}
-.main_layer #city_name {
-    padding: 8rem 1rem;
-    color: #a6a6b8;
-    font-weight: 300;
-    font-size: 2rem;
-}
-.main_layer .middle_layer{
-    display: flex;
-    justify-content: space-around;
-    color: #fff;
-}
-.main_layer .middle_layer p,
-.main_layer .middle_layer .fa{
-    font-size: 7rem;
-    font-weight: 900;
-}
-.fa{
-    color: #050f13;
-}
-sup{
-    font-size: 6rem;
-    margin-top: -2rem;
-}
-.data_hide{
-    visibility: hidden;
-    transition: all 0.3s linear;
-}
-.eroor img{
-    width: 60rem;
-    height: 30rem;
-}
-
-footer{
-   width: 100%;
-   background-color: var(--main-bg);
-   margin-top: auto;
-}
-footer p{
-    margin: 0;
-    text-align: center;
-    color: #fff;
-    line-height: 5rem;
-    font-size: 1.4rem;    
-}
-
-@media(max-width: 768px){
-    .main_header_left{
-        height: auto;
-        margin: 5rem 0;
-    }
-    .main_header_right{
-        height: auto;
-        margin-top: 3rem;
-    }
-}
-
-/* 404 error page style  */
