--- public/Projects/Anubhav/Weather App/templates/partials/footer.hbs
@@ -1,6 +1,0 @@
-<footer>
-    <p>
-        Created with ❤️ by Anubhav
-    </p>
-
-</footer>
