--- public/Projects/Anubhav/Weather App/templates/partials/navbar.hbs
@@ -1,32 +1,0 @@
-    <div class="container-fluid main_menu">
-        <div class="row">
-            <div class="col-md-10 col-12 mx-auto">
-                <nav class="navbar navbar-expand-lg">
-                    <a class="navbar-brand" href="#"><i class="fas fa-cloud-meatball"></i> Weather Website <i class="fas fa-cloud-meatball"></i></a>
-                    <button class="navbar-toggler" type="button" data-toggle="collapse"
-                        data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
-                        aria-expanded="false" aria-label="Toggle navigation">
-                        <span class="navbar-toggler-icon"></span>
-                    </button>
-
-                    <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                        <ul class="navbar-nav ml-auto">
-                            <li class="nav-item active">
-                                <a class="nav-link" href="/">Home <span class="sr-only">(current)</span></a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="/weather">Weather</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="/about">About Me</a>
-                            </li>
-                            
-                        </ul>
-                        
-                    </div>
-                </nav>
-
-            </div>
-        </div>
-
-    </div>
