--- public/Projects/Anubhav/Weather App/templates/views/404error.hbs
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  {{>header}}
-</head>
-
-<body>
-    {{>navbar}}
-
-    <div class="container-fluid main_header">
-        <div class="row">
-            <div class="col-md-10 col-12 mx-auto">
-                <div class="row">
-                    <!-- Left Side -->
-                    <div class="col-md-12 col-12 main_header_left align-items-center">
-                        <figure class = "eroor">
-                            <img src="images/1.png" alt="anubhav-image" class = "img-fluid" title="anubhav-img">
-                        </figure>
-                    
-                       <p>Weather Web App!!</p>
-                       <h1>{{errormsg}}</h1>
-                       <a href="/" target="_blank"><button>Go Back</button></a>
-                   </div>
-                    <!--Right side  -->
-                   
-                </div>
-            </div>
-        </div>
-    </div>   
-    
-    {{>footer}}
-</body>
-
-</html>
