--- public/Projects/Anubhav/Weather App/templates/views/about.hbs
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  {{>header}}
-</head>
-
-<body>
-    {{>navbar}}
-
-    <div class="container-fluid main_header">
-        <div class="row">
-            <div class="col-md-10 col-12 mx-auto">
-                <div class="row">
-                    <!-- Left Side -->
-                    <div class="col-md-6 col-12 main_header_right align-items-center">
-                        <figure>
-                            <img src="images/9.png" alt="anubhav-image" class = "img-fluid" title="anubhav-img">
-                        </figure>
-                    </div>    
-                   <div class = "col-md-6 col-12 main_header_left">
-                       <p>Weather Web App!!</p>
-                       <h1>I'm a <span class = "text_clr">Open Source and Tech Enthusiast</span></h1>
-                       <a href="https://drive.google.com/file/d/1GXHI7yajhkBQgRwd8Gs2p_5YMG3zStWr/view?usp=sharing" target="_blank"><button>About Me</button></a>
-                   </div>
-                    <!--Right side  -->
-                   
-                </div>
-            </div>
-        </div>
-    </div>   
-    
-    {{>footer}}
-</body>
-
-</html>
